Get_Cuptions/requirements.txt
youtube_transcript_apisqlalchemypsycopg2flaskflask-cors-3.0.10